REFACTOR refactors and cleans up client and server code
CLEANUP deletes NextPage module

CLEANUP deletes superfluous code

CLEANUP reorganizes code

REFACTOR refactors ProfilesStore logic

REFACTOR refactors HTML tags

REFACTOR refactors code

REFACTOR consolidates /coders and /codersNextPage routes

CLEANUP deletes /CountryCountByLanguage route

REFACTOR refactors odesk logic
