+++ client/app/stores/ProfilesStore.js
@@ -16,33 +16,19 @@ var page = 0;
 
 var ProfilesStore = assign({}, EventEmitter.prototype, {
   
-  // this function grabs the next page of profiles. TODO: rename function
-  getODeskData: function(page, language, country, subcategory, hourlyRateMax, minScore, maxScore) {
-    
-    var url =  'api/1/codersNextPage?page=' + page + '=&language=' + language + '=&country=' + country + '=&subcategory=' + subcategory + '=&hourlyRateMax=' + hourlyRateMax + '=&minScore=' + minScore + '=&maxScore=' + maxScore;
-    $.ajax({
-      url: url,
-      dataType: 'json',
-      type: 'GET',
-      success: function(data) {
-        profileData = data;
-        this.emitChange();
-      }.bind(this),
-      error: function(xhr, status, err) {
-        console.error('api/1/getAllFiles', status, err.toString());
-      }.bind(this)
-    });
+  getProfileDataFromServer: function(language, country, subcategory, hourlyRateMax, minScore, maxScore, page) {
+    return this.getCoders(language,country, subcategory, hourlyRateMax, minScore, maxScore, page);
   },
 
-  getCoders: function(language,country,subcategory, hourlyRateMax, minScore, maxScore) {
-    console.log("invoking getCoders");  
-    var url =  'api/1/coders?page=0=&language=' + language + '=&country=' + country + '=&subcategory=' + subcategory + '=&hourlyRateMax=' + hourlyRateMax + '=&minScore=' + minScore + '=&maxScore=' + maxScore;
+  getCoders: function(language,country,subcategory, hourlyRateMax, minScore, maxScore, page) {
     selectedLanguage = language;
     selectedCountry = country; 
     selectedSubcategory = subcategory;
     selectedHourlyRateMax = hourlyRateMax;
     selectedMinScore = minScore;
     selectedMaxScore = maxScore;
+    page = page || 0;
+    var url =  'api/1/coders?page=' + page + '=&language=' + language + '=&country=' + country + '=&subcategory=' + subcategory + '=&hourlyRateMax=' + hourlyRateMax + '=&minScore=' + minScore + '=&maxScore=' + maxScore;
 
     $.ajax({
       url: url,
@@ -60,14 +46,6 @@ var ProfilesStore = assign({}, EventEmitter.prototype, {
 
   },
 
-  getNextPageFromServer: function(page, language, country, subcategory, minScore, maxScore) {
-    return this.getODeskData(page, language, country, subcategory, minScore, maxScore);
-  },
-
-  getProfileDataFromServer: function(language, country, subcategory, hourlyRateMax, minScore, maxScore) {
-    return this.getCoders(language,country, subcategory, hourlyRateMax, minScore, maxScore);
-  },
-
   getProfileDataFromStore: function() {
     return profileData;
   },
@@ -76,18 +54,10 @@ var ProfilesStore = assign({}, EventEmitter.prototype, {
     this.emit(CHANGE_EVENT);
   },
 
-  /**
-   * @param {function} callback
-   */
-   //outgoing callbacks/changes
-
   addChangeListener: function(callback) {
     this.on(CHANGE_EVENT, callback);
   },
 
-  /**
-   * @param {function} callback
-   */
   removeChangeListener: function(callback) {
     this.removeListener(CHANGE_EVENT, callback);
   },
@@ -95,25 +65,22 @@ var ProfilesStore = assign({}, EventEmitter.prototype, {
   // Register callback to handle all updates
   dispatcherIndex: AppDispatcher.register(function(payload) {
     var action = payload.action;
-    var page = action.page || 0;
     var language = action.language || 'JavaScript';
     var country = action.country || 'Thailand';
     var subcategory = action.subcategory || 'Web Development';
     var hourlyRateMax = action.hourlyRateMax || '100';
     var minScore = action.minScore || 0;
     var maxScore = action.maxScore || 5;
+    var page = action.page || 0;
 
-    console.log(action.actionType);
     //incoming callbacks/changes
     switch(action.actionType) {
-      case 'PROFILES_NEXT_PAGE':
-        console.log('invoking PROFILES_NEXT_PAGE');
-        ProfilesStore.getNextPageFromServer(page, selectedLanguage, selectedCountry, selectedSubcategory, selectedHourlyRateMax, selectedMinScore, selectedMaxScore);
+      case 'GET_CODERS':
+        ProfilesStore.getProfileDataFromServer(language, country, subcategory, hourlyRateMax, minScore, maxScore);
         ProfilesStore.emitChange();
         break;
-      case 'GET_CODERS':
-        console.log('invoking GET_CODERS');
-        ProfilesStore.getProfileDataFromServer(language, country, subcategory, hourlyRateMax, minScore, maxScore);
+      case 'PROFILES_NEXT_PAGE':
+        ProfilesStore.getProfileDataFromServer(selectedLanguage, selectedCountry, selectedSubcategory, selectedHourlyRateMax, selectedMinScore, selectedMaxScore, page);
         ProfilesStore.emitChange();
         break;
     };
@@ -123,6 +90,5 @@ var ProfilesStore = assign({}, EventEmitter.prototype, {
 
   })
 });
-
 
 module.exports = ProfilesStore;
