--- server/api/1/freelancersLogic.js
@@ -1,10 +1,0 @@
-var http = require('http');
-// var config = require('./config.js');
-
-module.exports = {
-	// function getOdeskData(){
-
-	// } 
-};
-
-
